requirements.txt
pydantic~=2.11.7
fastapi~=0.115.13
pandas~=2.3.0
prophet~=1.1.7
python-dotenv~=1.1.1
py~=1.11.0
colorama~=0.4.6
lxml~=5.4.0
bs4~=0.0.2
beautifulsoup4~=4.13.4
soupsieve~=2.7
packaging~=25.0
html5lib~=1.1
h11~=0.16.0
ipython~=8.18.1
numpy~=2.0.2
pillow~=11.2.1
pip~=25.1.1
attrs~=25.3.0
wheel~=0.45.1
retry~=0.9.2
PySocks~=1.7.1
protobuf~=6.31.1
pyzmq~=27.0.0
cffi~=1.17.1
tornado~=6.5.1
setuptools~=80.3.1
comm~=0.2.2
traitlets~=5.14.3
parso~=0.8.4
jedi~=0.19.2
pytz~=2025.2
rich~=14.0.0
pygments~=2.19.2
markdown-it-py~=3.0.0
requests~=2.32.4
trio~=0.30.0
outcome~=1.3.0.post0
sniffio~=1.3.1
exceptiongroup~=1.3.0
sortedcontainers~=2.4.0
idna~=3.10
multidict~=6.5.1
propcache~=0.3.2
zipp~=3.23.0
MarkupSafe~=3.0.2
werkzeug~=3.1.3
click~=8.1.8
jinja2~=3.1.6
blinker~=1.9.0
itsdangerous~=2.2.0
pywin32~=310
jieba~=0.42.1
mdurl~=0.1.2
charset-normalizer~=3.4.2
patsy~=1.0.1
decorator~=5.2.1
scipy~=1.13.1
threadpoolctl~=3.6.0
matplotlib~=3.9.4
platformdirs~=4.3.8
akshare~=1.17.8
pyexecjs~=1.5.1
six~=1.17.0
joblib~=1.5.1
psutil~=7.0.0
xlrd~=2.0.2
openpyxl~=3.1.5
python-dateutil~=2.9.0.post0
plotly~=6.1.2
narwhals~=1.44.0
pydash~=7.0.7
multitasking~=0.0.11
func-timeout~=4.3.5
tqdm~=4.67.1
qstock~=1.3.8
jsonpath~=0.82.2
seaborn~=0.13.2
pyecharts~=2.0.8
backtrader~=1.9.78.123
pyfolio~=0.9.2
pywencai~=0.13.1
yarl~=1.20.1
async-timeout~=5.0.1
aiohttp~=3.12.13
aiosignal~=1.3.2
frozenlist~=1.7.0
aiohappyeyeballs~=2.6.1
urllib3~=2.5.0
certifi~=2025.6.15
debugpy~=1.8.14
starlette~=0.46.2
websockets~=15.0.1
anyio~=4.9.0
stack-data~=0.6.3
matplotlib-inline~=0.1.7
jupyter-client~=8.6.3
prompt-toolkit~=3.0.51
wcwidth~=0.2.13
importlib-resources~=6.5.2
holidays~=0.75
peewee~=3.18.1
empyrical~=0.5.5
scikit-learn~=1.6.1
cycler~=0.12.1
statsmodels~=0.14.4
tushare~=1.4.21
websocket-client~=1.8.0
uvicorn~=0.34.3
wsproto~=1.2.0
efinance~=0.5.5.2
webencodings~=0.5.1
et-xmlfile~=2.0.0
annotated-types~=0.7.0
simplejson~=3.20.1
selenium~=4.21.0
trio-websocket~=0.12.2
curl-cffi~=0.11.4
frozendict~=2.4.6
yfinance~=0.2.63
asttokens~=3.0.0
stanio~=0.5.1
cmdstanpy~=1.2.5
contourpy~=1.3.0
pandas-datareader~=0.10.0
executing~=2.2.0
fonttools~=4.58.4
ipykernel~=6.29.5
jupyter-core~=5.8.1
flask~=3.1.1
pure-eval~=0.2.3
pycparser~=2.22
prettytable~=3.16.0
pyparsing~=3.2.3
kiwisolver~=1.4.7
tomli~=2.0.1
autocommand~=2.2.2
inflect~=7.3.1
more-itertools~=10.3.0
typeguard~=4.3.0
importlib-metadata~=8.7.0
fake-useragent~=1.5.1
snownlp~=0.12.3